+++ workspaces/ui/src/components/diff/review-diff/ReviewBatchSelect.js
@@ -1,8 +1,7 @@
 import React, { useContext, useState } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
-import { AllCapturesContext, AllCapturesStore } from '../v2/CaptureManagerPage';
-import { Code } from './ICopyRender';
+import { AllCapturesContext, AllCapturesStore } from './AllCapturesContext';
 import FilterListIcon from '@material-ui/icons/FilterList';
 import time from 'time-ago';
 import {
