+++ workspaces/ui/src/components/diff/review-diff/learn-api/LearnAPIPageInner.js
@@ -1,22 +1,10 @@
 import React, { useContext, useMemo, useState } from 'react';
-import { useHistory } from 'react-router-dom';
 import { useCaptureContext } from '../../../../contexts/CaptureContext';
-import { useBaseUrl } from '../../../../contexts/BaseUrlContext';
 import { getOrUndefined, JsonHelper } from '@useoptic/domain';
-import { DiffLoadingOverview } from '../../v2/LoadingNextDiff';
-import { NewUrlModal } from '../../v2/AddUrlModal';
 import classNames from 'classnames';
-import { Show } from '../../../shared/Show';
-import isEqual from 'lodash.isequal';
 import RemoveCircleIcon from '@material-ui/icons/RemoveCircle';
 import FilterListIcon from '@material-ui/icons/FilterList';
 import { DocDarkGrey, DocGrey } from '../../../docs/DocConstants';
-import List from '@material-ui/core/List';
-import { PathNameFromId } from '../../../../contexts/EndpointContext';
-import ListItem from '@material-ui/core/ListItem';
-import { MethodRenderLarge, PathAndMethod } from '../../v2/PathAndMethod';
-import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
-import Chip from '@material-ui/core/Chip';
 import { lighten, makeStyles } from '@material-ui/core/styles';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
@@ -34,15 +22,14 @@ import TableRow from '@material-ui/core/TableRow';
 import Fade from '@material-ui/core/Fade';
 import PathMatcher from '../../PathMatcher';
 import { LearnAPIPageContext, LearnAPIStore } from './LearnAPIPageContext';
-import LearnAPIMenu from './LearnAPIMenu';
 import TextField from '@material-ui/core/TextField';
 import IconButton from '@material-ui/core/IconButton';
 import { LightTooltip } from '../../../tooltips/LightTooltip';
-import InProgressFullScreen from './InProgressFullScreen';
 import { useDiffSession } from '../ReviewDiffSession';
-import ButtonBase from '@material-ui/core/ButtonBase';
 import Button from '@material-ui/core/Button';
 import Box from '@material-ui/core/Box';
+import { DiffLoadingOverview } from '../DiffLoading';
+import { MethodRenderLarge } from '../PathAndMethod';
 
 export function LearnAPIPageInner(props) {
   const { urls, undocumentedEndpoints } = props;
