+++ workspaces/ui/src/components/docs/DocsPage.js
@@ -11,7 +11,6 @@ import {
   EndpointsContext,
   EndpointsContextStore,
 } from '../../contexts/EndpointContext';
-import { PathAndMethodLarge, SquareChip } from '../diff/v2/PathAndMethod';
 import Typography from '@material-ui/core/Typography';
 import { DocDivider } from './DocConstants';
 import { DocParameter } from './DocParameter';
@@ -19,16 +18,19 @@ import { HeadingContribution, MarkdownContribution } from './DocContribution';
 import { DESCRIPTION, PURPOSE } from '../../ContributionKeys';
 import groupBy from 'lodash.groupby';
 import ContentTabs, { RequestTabsContextStore } from './ContentTabs';
-import { BreadcumbX } from '../diff/v2/DiffNewRegions';
-import { ShapeExpandedStore } from '../diff/v2/shape_viewers/ShapeRenderContext';
-import { ShapeOnlyViewer } from '../diff/v2/shape_viewers/ShapeOnlyShapeRows';
-import { ShapeBox } from '../diff/v2/DiffReviewExpanded';
+import { ShapeExpandedStore } from '../diff/review-diff/shape-viewers/ShapeRenderContext';
+import { ShapeOnlyViewer } from '../diff/review-diff/shape-viewers/ShapeOnlyShapeRows';
 import Paper from '@material-ui/core/Paper';
 import EmptyState from '../support/EmptyState';
 import { AddOpticLink, DocumentingYourApi } from '../support/Links';
 import { trackUserEvent } from '../../Analytics';
 import { UpdateContribution } from '@useoptic/analytics/lib/events/diffs';
 import ScrollIntoViewIfNeeded from 'react-scroll-into-view-if-needed';
+import { BreadcumbX, ShapeBox } from '../diff/review-diff/BreadcrumbX';
+import {
+  PathAndMethodLarge,
+  SquareChip,
+} from '../diff/review-diff/PathAndMethod';
 const useStyles = makeStyles((theme) => ({
   maxWidth: {
     width: '100%',
