+++ workspaces/ui/src/entrypoints/demo-sessions.js
@@ -16,7 +16,6 @@ import {
   ColorButton,
   SubtleBlueBackground,
 } from '../theme';
-import { subtabs } from '../components/diff/v2/CaptureManagerPage';
 
 const snackbarStyles = makeStyles({
   alert: {
@@ -127,11 +126,7 @@ export default function DemoSessions(props) {
             message: `Nice! Descriptions will stay attached to their endpoint/fields even when the specification changes!\n\nLet's check back and see if there are any other diffs to approve`,
             action: {
               text: 'Review all diffs',
-              href: `/demos/todo/diffs/example-session?tab=${
-                hasDocumentedAllEndpointDiffs
-                  ? subtabs.UNDOCUMENTED_URL
-                  : subtabs.ENDPOINT_DIFF
-              }`,
+              href: `/demos/todo/diffs/example-session`,
             },
           });
           break;
