+++ workspaces/ui/src/routes.js
@@ -1,10 +1,8 @@
 import React, { Suspense } from 'react';
 import { Redirect, Switch, Route } from 'react-router-dom';
 import { useRouterPaths } from './RouterPaths';
-import { CaptureManagerPage } from './components/diff/v2/CaptureManagerPage';
 import { DocsPage } from './components/docs/DocsPage';
 import Loading from './components/navigation/Loading';
-import { ApiPage } from './components/api-page/ApiPage';
 import { SetupPage } from './components/setup-page/SetupPage';
 import { ReviewDiffPage } from './components/diff/review-diff/ReviewDiffPage';
 import { FinalizeSummaryContextStore } from './components/diff/review-diff/FinalizeSummaryContext';
@@ -30,11 +28,6 @@ export function ApiRoutes(props) {
           {/*<Route strict path={routerPaths.dashboardRoot} component={ApiPage} />*/}
           <Route exact path={routerPaths.setup} component={SetupPage} />
           <Route path={routerPaths.docsRoot} component={DocsPage} />
-          <Route
-            strict
-            path={routerPaths.diffsRoot}
-            component={CaptureManagerPage}
-          />
           <Route
             strict
             path={routerPaths.reviewRoot}
